store/wizard: extract cloud provider types into sub-slice
The cloud provider types (AwsShareMethod, AzureHyperVGeneration,
GcpAccountType, V1ListSourceResponseItem, CloudProviderSlice) were
scattered across component files and inlined in WizardState. Moving
them to store/slices/wizard/cloud/types.ts keeps type ownership in
the store layer and removes upward imports from store into components.
