--- src/Components/CreateImageWizard/types.ts
@@ -1,7 +1,0 @@
-// The Sources API only defines a V1ListSourceResponseItem[] type
-export type V1ListSourceResponseItem = {
-  id?: string | undefined;
-  name?: string;
-  source_type_id?: string;
-  uid?: string;
-};
